Add files via upload
All the abilities are working now. Though the timed abilities will be activated forever : Changes need to be done to timed abilities.  The Monster has been added. The Monster chases Ash using the A* algorithm. Confusion added for Gastly. Sleep added for JigglyPuff. Invincibility added for Zoroark. One random kill added for Squirtle. The game is not multiplayer but it can be played twice only on the same device. Work needs to be done to animate things, add information  and make the interface smoother and more attractive to the eye.